refactor(api): rename userProfile route to user-profile
Brings apps/api/src/routes into compliance with the project
ls-lint convention and globs the nested workspace ignore entries
(node_modules, .turbo, .vite, ...) so the validator stays silent
on every package directory.

Refs #162
